[v2.3.x] prov/efa: Add unit tests for LRU AH eviction
This commit also moves AH and conn functions from efa_av to efa_ah
and efa_conn files to enable the unit test

(cherry picked from commit 945fe592482821147041c89f8f0bde982a42c97e)
